Replace ApiPose with InputPose
