+++ src/isar/apis/models/models.py
@@ -2,6 +2,7 @@ from typing import List, Optional, Union
 from uuid import UUID
 
 from pydantic import BaseModel
+from alitra import Pose, Position, Orientation, Frame
 
 
 class StepResponse(BaseModel):
@@ -20,10 +21,51 @@ class StartMissionResponse(BaseModel):
     tasks: List[TaskResponse]
 
 
-class ApiPose(BaseModel):
+# We need to specify our own position/orientation/pose classes that do not contain the "Frame" class
+# because of an bug in generating the OpenAPI specification:
+# https://github.com/tiangolo/fastapi/issues/1505
+# This does not happen if all the classes are 'BaseModel' classes, but the alitra models are 'dataclasses',
+# hence the conversion being done.
+class InputOrientation(BaseModel):
     x: float
     y: float
     z: float
-    roll: float
-    pitch: float
-    yaw: float
+    w: float
+    frame_name: str
+
+    def to_alitra_orientation(self) -> Orientation:
+        return Orientation(
+            x=self.x,
+            y=self.y,
+            z=self.z,
+            w=self.w,
+            frame=Frame(self.frame_name),
+        )
+
+
+class InputPosition(BaseModel):
+    x: float
+    y: float
+    z: float
+    frame_name: str
+
+    def to_alitra_position(self) -> Position:
+        return Position(
+            x=self.x,
+            y=self.y,
+            z=self.z,
+            frame=Frame(self.frame_name),
+        )
+
+
+class InputPose(BaseModel):
+    position: InputPosition
+    orientation: InputOrientation
+    frame_name: str
+
+    def to_alitra_pose(self) -> Pose:
+        return Pose(
+            position=self.position.to_alitra_position(),
+            orientation=self.orientation.to_alitra_orientation(),
+            frame=Frame(self.frame_name),
+        )
